komponent lahku ajada et oleks mugavam töötada
