+++ angular-fe/src/app/_views/homePageView/blocks/homePageView.line.component.ts
@@ -1,0 +1,9 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'homepage-line',
+  templateUrl: 'homePageView.line.html',
+})
+export class HomePageLineComponent {
+  @Input() type: number = 1;
+}
