+++ angular-fe/src/app/_views/homePageView/blocks/homePageView.line.html
