+++ angular-fe/src/app/_views/homePageView/homePageView.component.ts
@@ -18,481 +18,6 @@ import { Subscription } from 'rxjs';
 import { ActivatedRoute } from '@angular/router';
 
 @Component({
-  selector: 'homepage-line',
-  templateUrl: 'homePageView.line.html',
-})
-export class HomePageLineComponent {
-  @Input() type: number = 1;
-}
-
-@Component({
-  selector: 'homepage-navblock',
-  templateUrl: 'blocks/homePageView.navblock.html',
-})
-export class HomePageNavBlockComponent {
-  @Input() data;
-  @Input() title: string;
-  @Input() description: string;
-  @Input() theme: string;
-
-  @HostBinding('class') get hostClasses(): string {
-    return `theme--${this.theme}`;
-  }
-
-}
-
-@Component({
-  selector: 'homepage-articles',
-  templateUrl: 'blocks/homePageView.articles.html',
-})
-export class HomePageArticlesComponent implements OnChanges {
-  @Input() data: any[] = [];
-  @Input() theme: string;
-  @Input() line: number = 1;
-  private imageList: string[] = [
-    'homepage-articles-1.svg',
-    'homepage-articles-2.svg',
-    'homepage-articles-3.svg',
-  ];
-
-  @HostBinding('class') get hostClasses(): string {
-    return `theme--${this.theme}`;
-  }
-
-  ngOnChanges() {
-    if (this.data) {
-      this.assignImages();
-    }
-  }
-
-  private assignImages() {
-
-    if (this.theme === 'career') {
-      this.imageList = [
-        'homepage-articles-career-1.svg',
-      ];
-    }
-    let counter = 0;
-    this.data = this.data.map((item, index) => {
-      if (counter >= this.imageList.length) {
-        counter = 0;
-      }
-      const image = `/assets/img/${this.imageList[counter]}`;
-      counter = counter + 1;
-      return {
-        ...item,
-        image,
-      };
-    });
-  }
-}
-
-@Component({
-  selector: 'homepage-events',
-  templateUrl: 'blocks/homePageView.events.html',
-})
-export class HomePageEventsComponent implements OnInit {
-  @Input() data: any[] = [];
-  @Input() theme: string;
-  @Input() title: string;
-  @Input() description: string;
-  @Input() line: number = 1;
-
-  private eventsAmount = 2;
-
-  private imageList: string[] = [
-    'homePage-events-1.svg',
-    'homePage-events-2.svg',
-  ];
-
-  constructor(
-    private http: HttpClient,
-    private settings: SettingsService,
-    private translate: TranslateService,
-  ) {
-  }
-
-  @HostBinding('class') get hostClasses(): string {
-    return `theme--${this.theme}`;
-  }
-
-  ngOnInit() {
-    this.getData();
-  }
-
-  private assignImages() {
-    let counter = 0;
-    this.data = this.data.map((item, index) => {
-      if (counter >= this.imageList.length) {
-        counter = 0;
-      }
-      const image = `/assets/img/${this.imageList[counter]}`;
-      counter = counter + 1;
-      return {
-        ...item,
-        image: {
-          url: image,
-        },
-      };
-    });
-  }
-
-  private getAdditional(entities): void {
-    const variables = {
-      lang: 'ET',
-    };
-    const query = this.settings.query('teachingPageAdditionalEvents', variables);
-    const subscription = this.http.get(query).subscribe((response: any) => {
-      try {
-        this.data = [
-          ...this.parseEvents(entities),
-          ...this.parseEvents(response.data.nodeQuery.entities),
-        ].slice(0, this.eventsAmount);
-        this.assignImages();
-      } catch (err) {
-      }
-    });
-  }
-
-  private parseEvents(items) {
-    return items.map((item) => {
-      return {
-        title: item.entityLabel,
-        author: item.fieldOrganizer,
-        created: item.fieldEventMainDate.unix,
-        content: item.fieldDescriptionSummary,
-        location: item.fieldEventLocation ? item.fieldEventLocation.name : false,
-        link: {
-          title: this.translate.get('button.read_more'),
-          url: {
-            path: item.entityUrl.path,
-          },
-        },
-        image: {
-          url: 'http://htm.wiseman.ee/sites/default/files/2020-02/homepage-slides-1.svg',
-        },
-      };
-    });
-  }
-
-  private getData(): void {
-
-    const variables = {
-      lang: 'ET',
-    };
-    const query = this.settings.query('teachingPageEvents', variables);
-
-    const subscription = this.http.get(query).subscribe((response: any) => {
-      try {
-        if (response.data.nodeQuery.entities.length < this.eventsAmount) {
-          this.getAdditional(response.data.nodeQuery.entities);
-        } else {
-          this.data = this.parseEvents(response.data.nodeQuery.entities);
-          this.assignImages();
-        }
-      } catch (err) {
-      }
-      subscription.unsubscribe();
-    });
-  }
-}
-
-@Component({
-  selector: 'homepage-careerDevelopment',
-  templateUrl: 'blocks/homePageView.careerDevelopment.html',
-})
-export class HomePageCareerDevelopmentComponent implements OnInit {
-  @Input() title: string;
-  @Input() description: string;
-  @Input() url: string;
-  @Input() theme: string;
-  @Input() line: number = 3;
-  public data: any[] = [];
-
-  constructor(
-    private http: HttpClient,
-    private settings: SettingsService,
-  ) {
-  }
-
-  @HostBinding('class') get hostClasses(): string {
-    return `theme--${this.theme}`;
-  }
-
-  ngOnInit() {
-    if (this.url) {
-      this.getData();
-    }
-  }
-
-  private getData(): void {
-    const variables = {
-      path: this.url,
-    };
-    const query = this.settings.query('getArticle', variables);
-    const subscription = this.http.get(query).subscribe((response: any) => {
-      try {
-        const accordionData = response.data.route.entity.fieldAccordionSection;
-        this.data = accordionData.map((item) => {
-          const slug = item.entity.fieldAccordionTitle.toLowerCase()
-            .replace(/span/g, '')
-            .replace(/<a href=".+?>/g, '')
-            .replace(/<\/a>/g, '')
-            .replace(/ /g, '-')
-            .replace(/[^A-Za-z0-9üõöä]+/igm, '-');
-          return {
-            slug,
-            title: item.entity.fieldAccordionTitle,
-            path: this.url,
-          };
-        });
-      } catch (err) {
-      }
-      subscription.unsubscribe();
-    });
-  }
-}
-
-@Component({
-  selector: 'homepage-slides',
-  templateUrl: 'blocks/homePageView.slides.html',
-})
-export class HomePageSlidesComponent {
-  @Input() title: string;
-  @Input() data: [] = [];
-  @Input() theme: string;
-  @Input() line: number = 2;
-
-  @HostBinding('class') get hostClasses(): string {
-    return `theme--${this.theme}`;
-  }
-}
-
-@Component({
-  selector: 'homepage-topical',
-  templateUrl: 'blocks/homePageView.topical.html',
-})
-export class HomePageTopicalComponent implements OnInit, OnChanges {
-  @Input() data: string;
-  @Input() theme: string;
-  @Input() line: number = 2;
-  @Input() category: string;
-  public article: any = {
-    title: '',
-    path: '',
-  };
-
-  constructor(
-    private http: HttpClient,
-    private settings: SettingsService,
-  ) {
-  }
-
-  @HostBinding('class') get hostClasses(): string {
-    return `theme--${this.theme}`;
-  }
-
-  ngOnInit() {
-    this.getData();
-  }
-
-  ngOnChanges() {
-    this.getData();
-  }
-
-  private getData() {
-    if (!this.data) {
-      return false;
-    }
-    const variables = {
-      path: this.data,
-    };
-    const query = this.settings.query('newsSingel', variables);
-    const subscription = this.http.get(query).subscribe((response) => {
-      this.article = {
-        title: '',
-        ...FieldVaryService(response['data']['route']['entity']),
-        path: this.data,
-      };
-      subscription.unsubscribe();
-    });
-  }
-}
-
-@Component({
-  selector: 'homepage-study',
-  templateUrl: 'blocks/homePageView.study.html',
-})
-export class HomePageStudyComponent {
-  @Input() theme: string;
-  @Input() line: number = 3;
-
-  @HostBinding('class') get hostClasses(): string {
-    return `theme--${this.theme}`;
-  }
-}
-
-@Component({
-  selector: 'homepage-slogan',
-  templateUrl: 'blocks/homePageView.slogan.html',
-})
-export class HomePageSloganComponent {
-  @Input() title: string = '';
-  @Input() person: string;
-  @Input() company: string;
-  @Input() theme: string;
-  @Input() line: number = 2;
-
-  @HostBinding('class') get hostClasses(): string {
-    return `theme--${this.theme}`;
-  }
-}
-
-@Component({
-  selector: 'homepage-footer',
-  templateUrl: 'blocks/homePageView.footer.html',
-})
-export class HomePageFooterComponent implements OnDestroy, AfterViewInit {
-  @Input() data: {
-    links,
-    logos,
-    contacts: [],
-  };
-  @Input() theme: string;
-  @Input() line: number = 4;
-  public subscribedStatus: boolean = false;
-  public subscribedError: boolean = false;
-  public loading: boolean = false;
-  public formGroup: FormGroup = this.formBuilder.group({
-    email: ['', [Validators.required, Validators.email]],
-  });
-  @ViewChild('scrollTarget') public scrollTarget;
-  private lang: string = 'ET';
-  private subscriptions: Subscription[] = [];
-  private tags: string = '';
-
-  constructor(
-    public settings: SettingsService,
-    private formBuilder: FormBuilder,
-    private alertsService: AlertsService,
-    private translate: TranslateService,
-    private http: HttpClient,
-    private cdr: ChangeDetectorRef,
-    private modalService: ModalService,
-    private route: ActivatedRoute,
-  ) {
-  }
-
-  @HostBinding('class') get hostClasses(): string {
-    return `theme--${this.theme}`;
-  }
-
-  public resetView(): void {
-    this.subscribedError = false;
-    this.subscribedStatus = false;
-    this.alertsService.clear('newsletter-order');
-  }
-
-  public submit(): void {
-
-    const data = {
-      queryId: this.settings.queryID('newsletterSignup'),
-      variables: {
-        lang: this.lang,
-        email: this.formGroup.controls.email.value,
-        tags: this.tags,
-      },
-    };
-    const query = `${this.settings.query('newsletterSignup')}`;
-
-    this.alertsService.clear('newsletter-order');
-
-    if (this.formGroup.invalid) {
-      this.alertsService.error(this.translate.get('newsletter.valid_email'), 'newsletter-order');
-      window.setTimeout(() => {
-        this.scrollTarget.el.nativeElement.scrollIntoView({ behavior: 'smooth' });
-      },                100);
-    } else {
-      this.loading = true;
-      const subscription = this.http.post(query, data).subscribe(
-        (response) => {
-          this.subscribedStatus = true;
-          this.loading = false;
-        },
-        (data) => {
-          this.subscribedError = true;
-          this.loading = false;
-        });
-      this.subscriptions.push(subscription);
-    }
-  }
-
-  ngOnInit() {
-    this.getTags();
-  }
-
-  ngAfterViewInit(): void {
-    if (this.route.snapshot.queryParams['confirmsubscription']) {
-      this.subscriptionModal(this.route.snapshot.queryParams['confirmsubscription']);
-    } else if (this.route.snapshot.queryParams['unsubscribe']) {
-      this.unsubscriptionModal(this.route.snapshot.queryParams['unsubscribe']);
-    }
-    this.cdr.detectChanges();
-  }
-
-  ngOnDestroy() {
-    this.subscriptions.forEach((element) => {
-      element.unsubscribe();
-    });
-  }
-
-  private subscriptionModal(token: string): void {
-    this.modalService.toggle('subscribe');
-    const data = {
-      variables: { token },
-      queryId: this.settings.queryID('newsletterActivate'),
-    };
-
-    const query = `${this.settings.query('newsletterActivate')}`;
-    const subscribe = this.http.post(query, data).subscribe((response) => {
-    });
-    this.subscriptions.push(subscribe);
-  }
-
-  private unsubscriptionModal(token: string): void {
-    this.modalService.toggle('unsubscribe');
-    const data = {
-      queryId: this.settings.queryID('newsletterDeactivate'),
-      variables: { token },
-    };
-    const query = `${this.settings.query('newsletterDeactivate')}`;
-    const subscribe = this.http.post(query, data).subscribe((response) => {
-      subscribe.unsubscribe();
-    });
-    this.subscriptions.push(subscribe);
-  }
-
-  private getTags(): void {
-    const variables = {
-      lang: this.lang,
-    };
-
-    const path = this.settings.query('newsletterTags', variables);
-
-    const subscription = this.http.get(path).subscribe((response) => {
-      try {
-        this.tags = response['data'].taxonomyTermQuery.entities.map((item) => {
-          return item.entityId;
-        }).join(', ');
-      } catch (err) {
-      }
-
-    });
-    this.subscriptions.push(subscription);
-  }
-}
-
-@Component({
   selector: 'homepage',
   templateUrl: 'homePageView.template.html',
   styleUrls: ['homePageView.styles.scss'],
